feat(UI): compile log is in a new buffer
This commit now adds the compile log, in a new buffer.
Previously this was in the messages window, which makes no sense.
It isn't easy to view your log and it dissapears on a single keystroke.

Support for colors from the compile log are missing, so need to add
support for that.